separate db related code
